third_party/WebKit/LayoutTests/fast/canvas/canvas-2d-imageData-create-nonfinite-expected.txt

Test the argument bounds of canvas createImageData.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ctx.createImageData(Infinity, Infinity) threw exception NotSupportedError: Failed to execute 'createImageData' on 'CanvasRenderingContext2D': The source width is infinite..
PASS ctx.createImageData(Infinity, 10) threw exception NotSupportedError: Failed to execute 'createImageData' on 'CanvasRenderingContext2D': The source width is infinite..
PASS ctx.createImageData(-Infinity, 10) threw exception NotSupportedError: Failed to execute 'createImageData' on 'CanvasRenderingContext2D': The source width is infinite..
PASS ctx.createImageData(10, Infinity) threw exception NotSupportedError: Failed to execute 'createImageData' on 'CanvasRenderingContext2D': The source height is infinite..
PASS ctx.createImageData(10, -Infinity) threw exception NotSupportedError: Failed to execute 'createImageData' on 'CanvasRenderingContext2D': The source height is infinite..
PASS ctx.createImageData(NaN, 10) threw exception NotSupportedError: Failed to execute 'createImageData' on 'CanvasRenderingContext2D': The source width is not a number..
PASS ctx.createImageData(10, NaN) threw exception NotSupportedError: Failed to execute 'createImageData' on 'CanvasRenderingContext2D': The source height is not a number..
PASS successfullyParsed is true

TEST COMPLETE
